refactor(client): features + fixes
a refactor of the menu adding many features and fixing other issues with resolution etc

a lost of features added is here bellow, which will be added to the docs soon

- cursor enable disable per menu /page
- scroll down key press/mouse
- element select only with cursor on
- max visible Items per menu
- label position of each element or all.
- fixed Height for background image
- multiple sliders for description
- slider for label elements
- confirm cancel buttons / can be label or do what ever you want them to do
- fixes the container height with elements with different heights
- grid now works with also key press correctly or with cursor
- input built in no need for vorp-inputs
- press right mouse click to go back  or backspace key
- press enter to select or left mouse click.
- hide show radar
- open and close sound with params
- added price system for the ui like RDO
- multiple tick boxes in the description
- label tick boxes
- fully backwards compatible
- footer text
- divider enable/disable
- fixes resolution for 4k and up
and prob some other stuff i cant remember

supports floats for prices and for the sliders

its fully backwards compatible , to use features you need to enabled them when building menus. this way we dont break anything its up to you.